+++ securedrop/source.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from datetime import datetime
 from cStringIO import StringIO
 from flask import (request, render_template, session, redirect, url_for,
                    flash, abort, g, send_file, Markup, make_response)
@@ -11,11 +10,10 @@ import crypto_util
 from flask_babel import gettext
 from rm import srm
 import store
-from db import db_session, Submission, Reply, get_one_or_else
+from db import db_session, Reply, get_one_or_else
 from source_app import create_app
 from source_app.decorators import login_required
-from source_app.utils import (logged_in, valid_codename, async_genkey,
-                              normalize_timestamps)
+from source_app.utils import logged_in, valid_codename
 
 import logging
 # This module's logger is explicitly labeled so the correct logger is used,
@@ -23,78 +21,6 @@ import logging
 log = logging.getLogger('source')
 
 app = create_app()
-
-
-@app.route('/submit', methods=('POST',))
-@login_required
-def submit():
-    msg = request.form['msg']
-    fh = request.files['fh']
-
-    # Don't bother submitting anything if it was an "empty" submission. #878.
-    if not (msg or fh):
-        flash(gettext(
-            "You must enter a message or choose a file to submit."),
-              "error")
-        return redirect(url_for('main.lookup'))
-
-    fnames = []
-    journalist_filename = g.source.journalist_filename
-    first_submission = g.source.interaction_count == 0
-
-    if msg:
-        g.source.interaction_count += 1
-        fnames.append(
-            store.save_message_submission(
-                g.filesystem_id,
-                g.source.interaction_count,
-                journalist_filename,
-                msg))
-    if fh:
-        g.source.interaction_count += 1
-        fnames.append(
-            store.save_file_submission(
-                g.filesystem_id,
-                g.source.interaction_count,
-                journalist_filename,
-                fh.filename,
-                fh.stream))
-
-    if first_submission:
-        msg = render_template('first_submission_flashed_message.html')
-        flash(Markup(msg), "success")
-
-    else:
-        if msg and not fh:
-            html_contents = gettext('Thanks! We received your message.')
-        elif not msg and fh:
-            html_contents = gettext('Thanks! We received your document.')
-        else:
-            html_contents = gettext('Thanks! We received your message and '
-                                    'document.')
-
-        msg = render_template('next_submission_flashed_message.html',
-                              html_contents=html_contents)
-        flash(Markup(msg), "success")
-
-    for fname in fnames:
-        submission = Submission(g.source, fname)
-        db_session.add(submission)
-
-    if g.source.pending:
-        g.source.pending = False
-
-        # Generate a keypair now, if there's enough entropy (issue #303)
-        entropy_avail = int(
-            open('/proc/sys/kernel/random/entropy_avail').read())
-        if entropy_avail >= 2400:
-            async_genkey(g.filesystem_id, g.codename)
-
-    g.source.last_updated = datetime.utcnow()
-    db_session.commit()
-    normalize_timestamps(g.filesystem_id)
-
-    return redirect(url_for('main.lookup'))
 
 
 @app.route('/delete', methods=('POST',))
