+++ securedrop/source_templates/lookup.html
@@ -23,7 +23,7 @@
 <p class="explanation">{{ gettext('You can send a file, a message, or both.') }}</p>
 <hr class="no-line">
 
-<form id="upload" method="post" action="/submit" enctype="multipart/form-data" autocomplete="off">
+<form id="upload" method="post" action="{{ url_for('main.submit') }}" enctype="multipart/form-data" autocomplete="off">
   <input name="csrf_token" type="hidden" value="{{ csrf_token() }}">
   <div class="snippet">
     <div class="attachment grid-item center">
